refactor(vscode): dedupe pixel-diff between preview and history panels
Same playbook as #807 (parseBounds + diffModeBar dedup): the
`computeDiffStats` algorithm — async client-side pixel diff between
two base64 PNGs — was identical between `preview/diffOverlay.ts` and
`history/historyDiffView.ts`, with the same ~90 LOC body and the same
`DiffStats` discriminated union shape. Lift both into a new
`webview/shared/pixelDiff.ts`.

The two prior copies had drifted slightly in defensive style:
- The preview copy used non-null `getContext("2d")!` assertions.
- The history copy guarded with `if (!ctx) resolve({ error: ... })`.

The shared module keeps the explicit guards (history-style) — they
catch a real failure mode (canvas-2d unavailable in some sandboxed
embeddings) and resolve cleanly instead of throwing inside the load
handler.

The history copy's flagged TODO ("a future shared module under
`webview/shared/pixelDiff.ts` could collapse them once the component
story stabilises") is now addressed.

`preview/diffOverlay.ts` and `history/historyDiffView.ts` re-export
`DiffStats` so existing imports
(`import { DiffStats } from "./diffOverlay"` etc.) keep working
without ripple-edits.

`preview/diffOverlay.ts`: 341 → 244 lines (−97 LOC).
`history/historyDiffView.ts`: 379 → 266 lines (−113 LOC).
`webview/shared/pixelDiff.ts`: +121 lines (new file).

Net: −89 LOC across the webview tree.

## Test plan

- [x] `npm run compile` clean (host + webview)
- [x] `npm test` — 377/377 passing
- [x] `npm run format:check` clean
- [ ] Manual smoke:
  - Live preview: focus a card, click Diff vs HEAD / Diff vs main —
    pixel-diff stats label renders (identical / changed / sizes
    differ).
  - History: shift+click two rows + toolbar Diff — inline diff
    banner shows `bytes identical` / `pixels differ`.
  - History: per-row "Diff vs previous" / "Diff vs current" — diff
    expansion's stats label transitions from "computing…" to a
    populated label.
