+++ vscode-extension/src/webview/history/historyDiffView.ts
@@ -11,17 +11,16 @@
 // orchestration — message dispatch, filter / scope plumbing, the
 // initial render kickoff.
 //
-// The pixel-diff helper (`computeDiffStats`) is duplicated semantically
-// from the live preview panel's `diffOverlay.ts`. Keeping the two
-// copies for now since the History panel is webview-self-contained
-// (no Lit dependency); a future shared module under
-// `webview/shared/pixelDiff.ts` could collapse them once the
-// component story stabilises.
+// The pixel-diff helper lives in `webview/shared/pixelDiff.ts` — the
+// preview panel's diff overlay reaches for the same algorithm.
 
 import { buildDiffModeBar, type DiffMode } from "../shared/diffModeBar";
+import { computeDiffStats, type DiffStats } from "../shared/pixelDiff";
 import type { HistoryDiffSummary } from "../shared/types";
 import type { VsCodeApi } from "../shared/vscode";
 import { cssEscape } from "./historyData";
+
+export type { DiffStats };
 
 interface DiffPayload {
     leftLabel: string;
@@ -33,24 +32,6 @@ interface DiffPayload {
 interface PersistedHistoryState {
     diffMode?: DiffMode;
 }
-
-export type DiffStats =
-    | { error: string }
-    | {
-          sameSize: false;
-          leftW: number;
-          leftH: number;
-          rightW: number;
-          rightH: number;
-      }
-    | {
-          sameSize: true;
-          w: number;
-          h: number;
-          diffPx: number;
-          total: number;
-          percent: number;
-      };
 
 export interface HistoryDiffViewConfig {
     /** Untyped vscode handle — `fillDiff` casts the persisted state to
@@ -120,100 +101,6 @@ export function fillDiff(
     renderHistoryDiffMode(body, initialMode, payload);
     computeDiffStats(payload.leftImage, payload.rightImage).then((s) => {
         applyDiffStats(stats, s);
-    });
-}
-
-/**
- * Async client-side pixel diff between two base64 PNG images. Resolves
- * a typed `DiffStats` discriminated union — error / size mismatch /
- * pixel count + same-size dimensions. Doesn't reject: every failure
- * mode lands as a `{ error }` variant so callers don't have to wrap
- * the call in a try/catch.
- *
- * Duplicated from `webview/preview/diffOverlay.ts`; same algorithm.
- */
-export function computeDiffStats(
-    leftBase64: string,
-    rightBase64: string,
-): Promise<DiffStats> {
-    return new Promise<DiffStats>((resolve) => {
-        const left = new Image();
-        const right = new Image();
-        let loaded = 0;
-        const onErr = (): void => resolve({ error: "image failed to load" });
-        const onOk = (): void => {
-            if (++loaded < 2) return;
-            try {
-                if (
-                    left.naturalWidth !== right.naturalWidth ||
-                    left.naturalHeight !== right.naturalHeight
-                ) {
-                    resolve({
-                        sameSize: false,
-                        leftW: left.naturalWidth,
-                        leftH: left.naturalHeight,
-                        rightW: right.naturalWidth,
-                        rightH: right.naturalHeight,
-                    });
-                    return;
-                }
-                const w = left.naturalWidth;
-                const h = left.naturalHeight;
-                const c1 = document.createElement("canvas");
-                c1.width = w;
-                c1.height = h;
-                const ctx1 = c1.getContext("2d");
-                if (!ctx1) {
-                    resolve({ error: "canvas 2d context unavailable" });
-                    return;
-                }
-                ctx1.drawImage(left, 0, 0);
-                const d1 = ctx1.getImageData(0, 0, w, h).data;
-                const c2 = document.createElement("canvas");
-                c2.width = w;
-                c2.height = h;
-                const ctx2 = c2.getContext("2d");
-                if (!ctx2) {
-                    resolve({ error: "canvas 2d context unavailable" });
-                    return;
-                }
-                ctx2.drawImage(right, 0, 0);
-                const d2 = ctx2.getImageData(0, 0, w, h).data;
-                let diff = 0;
-                const len = d1.length;
-                for (let i = 0; i < len; i += 4) {
-                    if (
-                        d1[i] !== d2[i] ||
-                        d1[i + 1] !== d2[i + 1] ||
-                        d1[i + 2] !== d2[i + 2] ||
-                        d1[i + 3] !== d2[i + 3]
-                    )
-                        diff++;
-                }
-                const total = w * h;
-                resolve({
-                    sameSize: true,
-                    w,
-                    h,
-                    diffPx: diff,
-                    total,
-                    percent: total > 0 ? diff / total : 0,
-                });
-            } catch (err) {
-                resolve({
-                    error:
-                        err instanceof Error
-                            ? err.message
-                            : "stats unavailable",
-                });
-            }
-        };
-        left.onload = onOk;
-        left.onerror = onErr;
-        right.onload = onOk;
-        right.onerror = onErr;
-        left.src = "data:image/png;base64," + leftBase64;
-        right.src = "data:image/png;base64," + rightBase64;
     });
 }
